Move expression definitions to ast module
