+++ src/parse/ast/item.rs
@@ -1,0 +1,5 @@
+/// Item values
+///
+/// An `Item` is a declaration made in the root context of a program
+/// -- namely the imort item `use`, and declarations such as `class`,
+/// `enum`, `struct`.
